refactor: extract PluginComposer from index.ts (issue #833)
Move the 1238-line composition root out of src/index.ts into a new
src/plugin-composer.ts (createPluginComposer().compose(ctx)). index.ts
is now a ~19-line delegation wrapper that preserves the exact Plugin
export and type re-exports.

Pure structural refactor: no behavior change, all 1619 tests pass. The
cache-safety drift guard is repointed to plugin-composer.ts atomically,
and codemaps are updated to reflect the new composition root.

Deliberately does NOT add hook-registry.ts — the Plugin return type
already serves as the hook registry, and ComposerDependencies is the test
seam. A priority-based registry would add ~100 lines for zero behavioral
benefit and break the drift-guard regex.
